requirements.txt
fastapi==0.89.1
pytest==7.2.1
pytest-cov==4.0.0
pylint==2.16.2
black==23.1.0
jose==1.0.0
passlib==1.7.4
python-consul==1.1.0
uvicorn~=0.20.0
pydantic~=1.10.4
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.22.0
google~=3.0.0
requests~=2.28.2
python-dotenv==0.21.1
starlette~=0.22.0
pymongo==4.3.3
pyjwt==2.6.0
python-multipart==0.0.5
bcrypt==4.0.1
